update paths to moved user to member components
