+++ pages/member/create/bio.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import Head from 'next/head';
 
-import ContentSection from '../../../../../components/shared/ContentSection';
-import togglePage from '../../../../../utilities/togglePage';
+import ContentSection from '../../../components/shared/ContentSection';
+import togglePage from '../../../utilities/togglePage';
 
-import Header from '../../../../../components/User/Profile/Create/Header';
-import Bio from '../../../../../components/User/Profile/Create/Bio';
+import Header from '../../../components/Member/Profile/Create/Header';
+import Bio from '../../../components/Member/Profile/Create/Bio';
 
 const CallForSpeakers = () => {
   return (
